[API CHANGE!] Change DRAMSim2 Interface
The main motivation for this commit is to make the simulator bindings
more opaque. By creating a pure virtual DRAMSimInterface class that
MultiChannelMemorySystem implements, a front end simulator can just use
the DRAMSimInterface pointer without having to pull in any other files.

Another motivation is to be able to give other memory simulators an interface
to be able to link instead of DRAMSim2 without having to rewrite all of
the bindings.