Add new address mapper
The new address mapping code is much more flexible than the old mapping
code. It can take a set of fields from a string such as "bk:rk:rw:of"
and map the fields in that order. For a list of fields that need to be
present in the mapping string, see AddressMapping.cpp.

Additionally, this is a performance win because address mapping only
ever has to be performed once the transaction is constructed. From that
point on, all of its fields are cached. This is especially a win in the
TransactionQueue scanning code since it used to re-do the address
mapping for every entry on every cycle.