PrintMacros: define SHOW_SIM_OUTPUT as weak symbol
When linking against libdramsim, library users have to define
SHOW_SIM_OUTPUT in order to avoid linking errors.

Fix it by emitting a weak version of the symbol. In case of the
DRAMSim tool, this weak symbol is overridden by the (strong) symbol
in the TraceBasedSim object.

Library users can optionally override the weak symbol, which means
that

a) New users of the library just won't have to worry about this,
   since they won't see an error at link time. They'll still be
   able to enable output by defining a strong symbol.

b) Existing users of the library will probably have already set
   a strong symbol, which will override the weak symbol we define
   here, and thus everything will work as before.

Signed-off-by: Emilio G. Cota <cota@braap.org>