DRAMSim: export Ini configuration getters
External simulators have no way of fetching configuration values
from the Ini file. This is a limitation that forces these simulators
to directly read the config file (e.g. gem5 with its dramsim_wrapper
class), which is a bad idea because a) the file is read twice
b) the DRAMSim overrides and sanity checks are ignored.

To avoid these limitations, the appended exports getters for
configuration values. These values are read directly from the
internal data structures, so that the configuration file is
only read once, as it should be.

Signed-off-by: Emilio G. Cota <cota@braap.org>